+++ pose/src/pose.cpp
@@ -8,27 +8,27 @@ using namespace std;
 
 Pose::Pose()
 {
-  (*this) = Pose(Vector3d(0,0,0), Vector3d(0,0,0));  
+  (*this) = Pose(Vector3f(0,0,0), Vector3f(0,0,0));  
 }
 
-Pose::Pose(const Vector3d& angles)
+Pose::Pose(const Vector3f& angles)
 {
-  (*this) = Pose(angles, Vector3d(0,0,0));
+  (*this) = Pose(angles, Vector3f(0,0,0));
 }
 
 
-Pose::Pose(const Vector3d& angles, const Vector3d& t)
+Pose::Pose(const Vector3f& angles, const Vector3f& t)
 {
   
-  this->T = Translation<double,3>(t)
-            *AngleAxis<double>(angles[0], Vector3d::UnitZ())
-            *AngleAxis<double>(angles[1], Vector3d::UnitX())
-            *AngleAxis<double>(angles[2], Vector3d::UnitY());
+  this->T = Translation<float,3>(t)
+            *AngleAxis<float>(angles[0], Vector3f::UnitZ())
+            *AngleAxis<float>(angles[1], Vector3f::UnitX())
+            *AngleAxis<float>(angles[2], Vector3f::UnitY());
 
   this->T = this->T.inverse();  
 }
 
-Pose::Pose(const Transform<double, 3, Affine>& _T)
+Pose::Pose(const Transform<float, 3, Affine>& _T)
 {
   this->T = _T;
 }
@@ -36,28 +36,28 @@ Pose::Pose(const Transform<double, 3, Affine>& _T)
 Pose::~Pose()
 {}
 
-Eigen::MatrixXd Pose::getRotation()
+Eigen::MatrixXf Pose::getRotation()
 {
   return this->T.rotation();
 }
 
-Eigen::Vector3d Pose::getTranslation()
+Eigen::Vector3f Pose::getTranslation()
 {
   return this->T.translation();
 }
 
-bool Pose::setPosition(Eigen::Vector3d position)
+bool Pose::setPosition(Eigen::Vector3f position)
 {
   this->T.translation() = position;
   return true;
 }
 
 
-Pose Pose::getRotationAroundAxis(Vector3d axis, double angle)
+Pose Pose::getRotationAroundAxis(Vector3f axis, float angle)
 {
   Pose result;
 
-  result.T = AngleAxis<double>(angle, axis);
+  result.T = AngleAxis<float>(angle, axis);
   result.T = result.T.inverse();
 
   return result;
